On second thought I don't like passing an UUID around as a unique nonce.
Instead, generate a nonce that looks exactly the same as the old, but using
the standard random_bytes() PHP7 function. The paragonie/random_compat dependency
has been made explicit in the composer.json file,  but it was required
anyway by the Symfony Components 2.8 and up through the symfony/polyfill-php70
package, so it doesn't really count as a new dependency.